refactor: uselayoutstyles
